Use a vertical icon for the ellipsis menu
